refactorin some funtions utils
